Move touch events into streaming manager. Create callback for remotedisplay created.
